Add runebender-lib crate
This is the start of separating out the reuseable code from
the stuff that is specifically part of the desktop application.
