+++ .gitignore
@@ -1,2 +1,3 @@
 /target
 **/*.rs.bk
+**/.DS_Store
