+++ Cargo.toml
@@ -6,6 +6,7 @@ license = "Apache-2.0"
 edition = "2018"
 
 [dependencies]
+runebender-lib = { version = "0.1.0", path = "runebender-lib" }
 norad = { version = "0.3.1", features = ["druid", "rayon"] }
 druid = "0.7"
 log = "0.4.8"
@@ -22,4 +23,4 @@ druid-theme-loader = { version = "0.1.0", path = "druid-theme-loader", features 
 druid = { version = "0.7.0", git = "https://github.com/linebender/druid.git", rev = "fb434a1f" }
 
 [workspace]
-members = ["druid-theme-loader"]
+members = ["runebender-lib", "druid-theme-loader"]
