+++ runebender-lib/src/widgets/coord_pane.rs
